+++ spec/support/apps/rails_5.2_app/config/application.rb
@@ -5,8 +5,8 @@ require "active_record/railtie"
 
 Bundler.require(*Rails.groups)
 
-module Rails51App
+module Rails52App
   class Application < Rails::Application
-    config.load_defaults 5.1
+    config.load_defaults 5.2
   end
 end